入库单状态

+++ clms-warehouse-manage/warehouse-api/src/main/java/com/lcdt/warehouse/vo/ConstantVO.java
@@ -7,7 +7,7 @@ package com.lcdt.warehouse.vo;
 public class ConstantVO {
 
     //入库单状态
-    public static short IN_ORDER_STATUS_WATIE_STORAGE=1; //待入库
-    public static short IN_ORDER_STATUS_HAVE_STORAGE=2;//已入厂
-    public static short IN_ORDER_STATUS_HAVE_CANCEL=3;//已取消
+    public static int IN_ORDER_STATUS_WATIE_STORAGE=1; //待入库
+    public static int IN_ORDER_STATUS_HAVE_STORAGE=2;//已入厂
+    public static int IN_ORDER_STATUS_HAVE_CANCEL=3;//已取消
 }
